// components/Transactions/types.ts
export interface TransactionProps {
  id: number;
  title: string;
  subTitle: string;
  amount: string;
  date: string;
  art: {
    icon: string;
    background: string;
  };
}

export interface TransactionSectionProps {
  data: TransactionProps[];
}

export interface TransactionAviProps {
  icon: any;
  background: string;
}